requirements.txt
cassandra-driver==3.24.0
click==7.1.2
Flask==1.1.2
geomet==0.2.1.post1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
six==1.15.0
Werkzeug==1.0.1
pytest==6.1.1
